separated modules, implement PartialEq for Matrix (for matching data payload)
